src/dedupbyani/requirements.txt
pyfastani ~=0.1.1
unionfind
biopython >=1.79
